Add Hevea specific instructions
Hevea is used to generate an HTML version and it supports conditional
compiling. To maximize portability, Hevea comments are used instead of
the `hevea` package.

Hevea does not support the wrapfig environment, so Hevea is told to
ignore it.

Hevea does not support `\hrule`, so a definition is given to Hevea.

The PDF is authoritative, so point to it in Hevea's render.